updated: 1. 优化了部分代码； 2. 新增了Client::channel()及配置channel.php，支持多通道；

+++ src/Process/AbstractProcess.php
@@ -20,7 +20,7 @@ abstract class AbstractProcess
     public function __construct()
     {
         $this->client = new NacosClient();
-        $this->retry_interval = config('plugin.workbunny.webman-nacos.app.process_retry_interval', 5);
+        $this->retry_interval = (int)config('plugin.workbunny.webman-nacos.app.process_retry_interval', 5);
     }
 
     /**
